fix: addressed tx id page feedback
