+++ src/app/txid/[txId]/redesign/CoinbasePage.tsx
@@ -1,4 +1,5 @@
-import { Grid } from '@chakra-ui/react';
+import { Text } from '@/ui/Text';
+import { Grid, Stack } from '@chakra-ui/react';
 
 import {
   CoinbaseTransaction,
@@ -8,16 +9,24 @@ import {
 
 import { DetailsCard } from './DetailsCard';
 import { TxHeader } from './TxHeader';
-import { TxSummary } from './TxSummary';
+import { TabsContentContainer } from './TxTabs';
+import { TxSummary } from './tx-summary/TxSummary';
 
 export const CoinbasePage = ({ tx }: { tx: CoinbaseTransaction | MempoolCoinbaseTransaction }) => {
   return (
     <>
       <TxHeader tx={tx} />
-      <Grid templateColumns={{ base: '1fr', md: '1fr auto' }} gap={2}>
-        <TxSummary tx={tx} />
-        <DetailsCard tx={tx as Transaction} />
-      </Grid>
+      <Stack gap={3}>
+        <Text textStyle="heading-xs">Overview</Text>
+
+        <Grid templateColumns={{ base: '1fr', md: '75% 25%' }} gap={2}>
+          <TabsContentContainer>
+            <TxSummary tx={tx} />
+          </TabsContentContainer>
+
+          <DetailsCard tx={tx as Transaction} />
+        </Grid>
+      </Stack>
     </>
   );
 };
