+++ src/app/txid/[txId]/redesign/DetailsCard.tsx
@@ -9,7 +9,7 @@ import { Stack } from '@chakra-ui/react';
 import { CoinbaseTransaction, Transaction } from '@stacks/stacks-blockchain-api-types';
 
 import { useTxBlock } from '../useTxBlock';
-import { SummaryItemLabel, SummaryItemValue } from './TxSummary';
+import { SummaryItemLabel, SummaryItemValue } from './tx-summary/TxSummary';
 
 export function DetailsCard({ tx }: { tx: Transaction }) {
   const items = getDetailsCardItems(tx);
@@ -24,8 +24,6 @@ export function DetailsCard({ tx }: { tx: Transaction }) {
       borderColor="redesignBorderSecondary"
       gap={4}
       h="fit-content"
-      minWidth="250px"
-      maxWidth="350px"
     >
       <Text textStyle="text-medium-sm" color="textPrimary">
         {title}
@@ -36,7 +34,7 @@ export function DetailsCard({ tx }: { tx: Transaction }) {
 }
 
 function getDetailsCardTitle(tx: Transaction) {
-  if (tx.tx_type === 'coinbase') return 'Bitcoin anchor details';
+  if (tx.tx_type === 'coinbase') return 'Burn block details';
   if (tx.tx_type === 'token_transfer') return null;
   if (tx.tx_type === 'smart_contract') return null;
   if (tx.tx_type === 'contract_call') return null;
@@ -85,7 +83,7 @@ function CoinbaseDetailsCardItems({ tx }: { tx: CoinbaseTransaction }) {
       <SummaryItem
         label="Block height"
         value={block?.burn_block_height?.toString()}
-        valueRenderer={value => <BlockHeightBadge blockType="stx" blockHeight={Number(value)} />}
+        valueRenderer={value => <BlockHeightBadge blockType="btc" blockHeight={Number(value)} />}
         showCopyButton={true}
       />
       <SummaryItem
