+++ src/app/txid/[txId]/redesign/Events.tsx
@@ -2,6 +2,7 @@ import { Stack } from '@chakra-ui/react';
 
 import { MempoolTransaction, Transaction } from '@stacks/stacks-blockchain-api-types';
 
+import { TabsContentContainer } from './TxTabs';
 import { EventsTableFilters } from './events-table/EventsTableFilters';
 import { EventsTableWithFilters } from './events-table/EventsTableWithFilters';
 import { EventsTableFiltersProvider } from './events-table/filters/useEventsTableFilters';
@@ -11,7 +12,9 @@ export function Events({ tx }: { tx: Transaction | MempoolTransaction }) {
     <EventsTableFiltersProvider>
       <Stack>
         <EventsTableFilters />
-        <EventsTableWithFilters txId={tx.tx_id} initialData={undefined} />
+        <TabsContentContainer>
+          <EventsTableWithFilters txId={tx.tx_id} initialData={undefined} />
+        </TabsContentContainer>
       </Stack>
     </EventsTableFiltersProvider>
   );
