+++ src/app/txid/[txId]/redesign/TxTabs.tsx
@@ -1,4 +1,3 @@
-import { ScrollIndicator } from '@/common/components/ScrollIndicator';
 import { ValueBasisFilterPopover } from '@/common/components/table/filters/value-basis-filter/ValueBasisFiterPopover';
 import { TabsContent, TabsList, TabsRoot, TabsTrigger } from '@/ui/Tabs';
 import { Text } from '@/ui/Text';
@@ -8,17 +7,19 @@ import { useState } from 'react';
 import { MempoolTransaction, Transaction } from '@stacks/stacks-blockchain-api-types';
 
 import { Events } from './Events';
-import { TxSummary } from './TxSummary';
 import { FunctionCalled } from './function-called/FunctionCalled';
+import { TxSummary } from './tx-summary/TxSummary';
 
 function TabTriggerComponent({
   label,
   value,
+  secondaryLabel,
   isActive,
   onClick,
 }: {
   label: string;
   value: string;
+  secondaryLabel?: string;
   isActive: boolean;
   onClick: () => void;
 }) {
@@ -37,13 +38,16 @@ function TabTriggerComponent({
       px={3}
       onClick={onClick}
     >
-      <Text
-        font="matterRegular"
-        textStyle="text-regular-xl"
-        color={isActive ? 'textPrimary' : 'textTertiary'}
-      >
-        {label}
-      </Text>
+      <Flex gap={1} alignItems="center">
+        <Text textStyle="heading-xs" color={isActive ? 'textPrimary' : 'textSecondary'}>
+          {label}
+        </Text>
+        {secondaryLabel && (
+          <Text textStyle="heading-xs" color={isActive ? 'textPrimary' : 'textTertiary'}>
+            {secondaryLabel}
+          </Text>
+        )}
+      </Flex>
     </TabsTrigger>
   );
 }
@@ -90,11 +94,12 @@ function getTabsTriggersByTransactionType(
           onClick={() => setSelectedTab(TransactionIdPageTab.Overview)}
         />
         <TabTriggerComponent
-          key={TransactionIdPageTab.Events}
-          label={`Events ${numTxEvents > 0 ? `(${numTxEvents})` : ''}`}
-          value={TransactionIdPageTab.Events}
-          isActive={selectedTab === TransactionIdPageTab.Events}
-          onClick={() => setSelectedTab(TransactionIdPageTab.Events)}
+          key="events"
+          label={`Events`}
+          secondaryLabel={numTxEvents > 0 ? `(${numTxEvents})` : ''}
+          value="events"
+          isActive={selectedTab === 'events'}
+          onClick={() => setSelectedTab('events')}
         />
       </>
     );
@@ -162,9 +167,7 @@ function getTabsContentByTransactionType(tx: Transaction | MempoolTransaction) {
           </TabsContentContainer>
         </TabsContent>
         <TabsContent key="events" value="events" w="100%">
-          <TabsContentContainer>
-            <Events tx={tx} />
-          </TabsContentContainer>
+          <Events tx={tx} />
         </TabsContent>
       </>
     );
@@ -212,7 +215,6 @@ function getTabsContentByTransactionType(tx: Transaction | MempoolTransaction) {
 
 export const TxTabs = ({ tx }: { tx: Transaction | MempoolTransaction }) => {
   const [selectedTab, setSelectedTab] = useState('overview');
-
   return (
     <TabsRoot
       variant="primary"
@@ -222,14 +224,16 @@ export const TxTabs = ({ tx }: { tx: Transaction | MempoolTransaction }) => {
       borderRadius="redesign.xl"
       w="full"
     >
-      <Flex justifyContent={'space-between'} w="full" gap={2}>
-        <ScrollIndicator>
-          <TabsList>{getTabsTriggersByTransactionType(tx, selectedTab, setSelectedTab)}</TabsList>
-        </ScrollIndicator>
-        <Flex alignItems={'center'} gap={4}>
-          <Text textStyle="text-regular-sm">Show:</Text>
-          <ValueBasisFilterPopover />
-        </Flex>
+      <Flex justifyContent={'space-between'} w="full">
+        <TabsList flexWrap={'wrap'}>
+          {getTabsTriggersByTransactionType(tx, selectedTab, setSelectedTab)}
+        </TabsList>
+        {tx.tx_type === 'token_transfer' && (
+          <Flex alignItems={'center'} gap={2}>
+            <Text textStyle="text-regular-sm">Show:</Text>
+            <ValueBasisFilterPopover />
+          </Flex>
+        )}
       </Flex>
       {getTabsContentByTransactionType(tx)}
     </TabsRoot>
